+++ lib/decorator/Crossover.ts
@@ -5,5 +5,7 @@ const KEY = 'Crossover';
 export const Crossover = () => {
     return (target: any, key: string, descriptor: any) => {
         StorageInstance.set(KEY, descriptor.value);
+
+        return descriptor;
     };
 };