+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## 0.1.2 (1.02.2018)
+* fix: decorators context
+* removed: mutation rate param
+
 ## 0.1.1 (26.01.2018)
 * fix: package.json